% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_anonymization.R
\name{attach_anon_id}
\alias{attach_anon_id}
\title{Add anonymized site identifier to all tables}
\usage{
attach_anon_id(all_sites_tbl, tbls_to_anon)
}
\arguments{
\item{all_sites_tbl}{a tbl that contains a column called \code{site}
with each of the sites that you want to mask
in all the other tables}

\item{tbls_to_anon}{a \strong{NAMED} list of all of the tables that you want to add
the masked site identifier column to}
}
\value{
each of the original tables in \code{tbls_to_anon} with the original columns
plus a column called \code{site_anon} with a masked identifier
that is consistent across all of the tables in \code{tbls_to_anon}
}
\description{
Add anonymized site identifier to all tables
}
