% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{bmc_assign}
\alias{bmc_assign}
\title{assign "best" "not best" labels to concepts}
\usage{
bmc_assign(bmc_output, conceptset)
}
\arguments{
\item{bmc_output}{the bmc_counts output from check_bmc}

\item{conceptset}{the bmc_concepts output from check_bmc with an additional
column called "include" added with "not best" or unideal concepts
marked with a 0}
}
\value{
a dataframe with the include designations attached to all concepts; also checks
to ensure no additional values are present
}
\description{
assign "best" "not best" labels to concepts
}
\keyword{internal}
