% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmc_best_notbest}
\alias{bmc_best_notbest}
\title{Best Mapped Concepts Best/Not Best File}
\format{
\subsection{\code{bmc_best_notbest}}{

A dataframe or CSV file with 4 columns
\describe{
\item{check_name}{The name of the check associated with the concepts of interest, which should match what is provided in the \code{bmc_tbl} input}
\item{concept}{The concept identifier or concept name (depending on if a vocabulary table is provided) that should be classified as best/not best}
\item{best_notbest}{A binary indicator where 0 indicates the concept is not best / unacceptable and 1 indicates the concept is best / acceptable}
\item{default_to}{An indicator, either \code{best} or \code{notbest}, to designate what any concepts not listed in this file should be defaulted to}
}
}
}
\usage{
bmc_best_notbest
}
\description{
A sample version of the best/not best designation file to classify
concepts in the result of this module. This exact
file is also included as a CSV in the package if the user wishes
to use it, or the structure can be copied to produce a custom
list of checks.
}
\keyword{datasets}
