% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{bmc_rollup}
\alias{bmc_rollup}
\title{Function to compute proportion of "best" based on output from bmc check}
\usage{
bmc_rollup(bmc_output_pp)
}
\arguments{
\item{bmc_output_pp}{table output from the bmc_assign function, which has all the
columns output from the bmc check + an indicator column for whether the concept
should be in the "best" category}
}
\value{
table with the cols: site, check_type, database_version, check_name,
check_desc,  count_best, include, total_rows, total_pts, best_row_prop, best_pts_prop
}
\description{
Function to compute proportion of "best" based on output from bmc check
}
