% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{check_bmc}
\alias{check_bmc}
\title{Best Mapped Concepts}
\usage{
check_bmc(
  bmc_tbl,
  best_notbest_tbl,
  omop_or_pcornet = "omop",
  concept_tbl = NULL,
  check_string = "bmc"
)
}
\arguments{
\item{bmc_tbl}{\emph{tabular input} || \strong{required}

The primary input table that contains descriptive information about the checks
to be executed by the function. It should include definitions the fields that should be
evaluated to determine if they only include "best" concepts.
see \code{?bmc_input_omop} or \code{?bmc_input_pcornet} for examples of the input structure}

\item{best_notbest_tbl}{\emph{tabular input} || \strong{required}

A table indicating the best/not best concept designations for each
check. See \code{?bmc_best_notbest} for an example of this input structure.

If it is easier to list the concepts that are best, like if there
are a limited number of acceptable concepts but many unacceptable concepts,
the \code{best_notbest} column should be set to \code{1} and the \code{default_to} column
should be set to \code{notbest}. If the inverse is true and it is easier to
list the unacceptable concepts, the \code{best_notbest} column should be
set to \code{0} and the \code{default_to} column should be set to \code{best}.}

\item{omop_or_pcornet}{\emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{concept_tbl}{\emph{tabular input} || defaults to \code{NULL}

An optional parameter used to define a vocabulary table with concept definitions
(for example, the OHDSI concept table). If left NULL, the concepts as they exist
in the fact table will be returned to the user.}

\item{check_string}{\emph{string} || defaults to \code{bmc}

An abbreviated identifier that will be used to label all output from this module}
}
\value{
This function will return a table with one row for each concept present
in each user-defined field, the associated row and patient counts/proportions,
and a label indicating whether the concept is considered "best" or "not best"
}
\description{
This function will identify the existing concepts within the specified field
so the user can assess which of these concepts are acceptable ("best") or should
not be used in that field ("not best")
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::bmc_input_omop
ndq::bmc_input_pcornet

# Then, create a table that indicates specific concepts that you
# want to consider "best" or "not best," and indicate your preference
# for whether other concepts should default to one or the other.
ndq::bmc_best_notbest

# Use this as your input to the BMC function
\dontrun{
my_bmc_rslt <- check_bmc(bmc_tbl = ndq::bmc_input_omop,
                         best_notbest_tbl = ndq::bmc_best_notbest,
                         omop_or_pcornet = 'omop',
                         concept_tbl = vocabulary_tbl("concept"), ## points to OHDSI concept table
                         check_string = 'bmc')
}


}
