% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_fact_documentation.R
\name{check_cfd}
\alias{check_cfd}
\title{Clinical Fact Documentation}
\usage{
check_cfd(
  cfd_tbl,
  visit_type_filter,
  visit_type_tbl,
  omop_or_pcornet = "omop",
  visit_tbl = cdm_tbl("visit_occurrence"),
  check_string = "cfd"
)
}
\arguments{
\item{cfd_tbl}{\emph{tabular input} || \strong{required}

The primary input table that contains descriptive information about the checks
to be executed by the function. It should include definitions for the clinical fact
types that should be evaluated against the visit_tbl.
see \code{?cfd_input_omop} or \code{?cfd_input_pcornet} for examples of the input structure}

\item{visit_type_filter}{\emph{string / vector} || \strong{required}

A string or vector of strings that specifies the visit type(s) the function
should limit to for the analysis (i.e. \code{inpatient}, \code{c(inpatient, outpatient)}).
If \code{all} is included as a visit type, all available visit types will be pulled
from the \code{visit_tbl} (not just what was defined in the \code{visit_type_tbl}) to capture
the full array of visits.}

\item{visit_type_tbl}{\emph{tabular input} || \strong{required}

A table with mappings that link visit_concept_ids / enc_types to a broader
string descriptor (i.e. inpatient). Multiple rows should be included if
multiple IDs should be mapped to the same string label.}

\item{omop_or_pcornet}{\emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{visit_tbl}{\emph{tabular input} || defaults to \code{cdm_tbl('visit_occurrence')}

The CDM table with the visits to be used in the analysis. Typically, this will
be either the OMOP \code{visit_occurrence} or PCORnet \code{encounter} table}

\item{check_string}{\emph{string} || defaults to \code{cfd}

An abbreviated identifier that will be used to label all output from this module}
}
\value{
A summary table that, for each provided clinical fact and visit type combination,
will include the count and proportion of visits that are and are not linked to
at least one occurrence of the fact type of interest. It will also include the count
and proportion of patients that are associated with these visits.
}
\description{
This function will identify visits that both do and do not link to at least one occurrence of the
user-specified clinical facts. It will also compute the counts of patients are associated with
these visits.
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::cfd_input_omop
ndq::cfd_input_pcornet

# Next define the visit types to be examined
# You can access examples for both OMOP & PCORnet here:
ndq::cfd_visit_types_omop
ndq::cfd_visit_types_pcornet

# Use this as your input to the CFD function
\dontrun{
my_cfd_rslt <- check_cfd(cfd_tbl = ndq::cfd_input_omop,
                         visit_type_tbl = ndq::cfd_visit_types_omop,
                         visit_type_filter = c('inpatient', 'outpatient'),
                         omop_or_pcornet = 'omop',
                         visit_tbl = cdm_tbl('visit_occurrence'),
                         check_string = 'cfd')
}

}
