% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cycle_changes.R
\name{check_dc}
\alias{check_dc}
\title{Data Cycle Changes}
\usage{
check_dc(
  dc_tbl,
  omop_or_pcornet = "omop",
  prev_db_string = "v1",
  current_db_string = "v2",
  prev_ct_src = "cdm",
  prev_db = config("db_src"),
  prev_rslt_tbl = "dc_output",
  prev_rslt_schema = config("results_schema"),
  check_string = "dc"
)
}
\arguments{
\item{dc_tbl}{table with details of the CDM elements that should be tested in
both the current and previous data model versions}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{prev_db_string}{string label indicating the previous CDM version; if
prev_ct_src == 'result', ensure this matches the appropriate database
label in the last set of results}

\item{current_db_string}{string label indicating the current CDM version}

\item{prev_ct_src}{a string indicating where the counts from the previous data model
should be extracted: either \code{cdm} (to pull from the previous CDM instance)
or \code{result} (to pull from a previous instance of check_dc output)}

\item{prev_db}{the database connection to be used to access the previous CDM or results;
defaults to \code{config('db_src')}}

\item{prev_rslt_tbl}{if prev_ct_src = 'result', the name of the table where previous results
are stored}

\item{prev_rslt_schema}{if prev_ct_src = 'result', the name of the schema where previous results
are stored. defaults to \code{config('results_schema')}}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{dc}}
}
\value{
a list with two dataframes:
- \code{dc_cts}: dataframe containing the row and (where applicable) person counts for each table
- \code{dc_meta}: the metadata associated with each input table that appears in \code{dc_cts}
}
\description{
Data Cycle Changes
}
