% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_concordance.R
\name{check_dcon}
\alias{check_dcon}
\title{Domain Concordance}
\usage{
check_dcon(
  dcon_tbl,
  compute_level = "patient",
  omop_or_pcornet = "omop",
  check_string = "dcon"
)
}
\arguments{
\item{dcon_tbl}{\emph{tabular} || \strong{required}

The primary input table that contains descriptive information about the checks
to be executed by the function. It should include definitions for each pair of cohorts,
meaning there should be \strong{2 rows per check}, one fo each cohort, with the same \code{check_id}.
see \code{?dcon_input_omop} or \code{?dcon_input_pcornet} for examples of the input structure}

\item{compute_level}{\emph{string} || defaults to \code{patient}

A string indicating whether the analysis should be conducted at the \code{patient},
\code{visit}, or \code{visit_detail} (for OMOP implementations) level}

\item{omop_or_pcornet}{\emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{check_string}{\emph{string} || defaults to \code{dcon}

An abbreviated identifier that will be used to label all output from this module}
}
\value{
This function will return a list of two dataframes:
\itemize{
\item \code{dcon_output}: A table with counts of patients/visits in the first cohort,
patients/visits in the second cohort, and patients/visits in both cohorts (who also
meet the time limit criteria)
\item \code{dcon_meta}: A table with descriptive metadata about each of the cohorts for
which counts were computed
}
}
\description{
Given the details of a pair of clinical events, this function
will determine the count of patients OR visits that meet criteria for
the first event, the second event, and both events. Users can optionally
define a time limitation for the combined cohort, so patients/visits will only
count towards that cohort when the two events occur within a specified number of days
of each other.
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::dcon_input_omop
ndq::dcon_input_pcornet

# Use this as your input to the DCON function
## To execute the check at the patient level:
\dontrun{
my_dcon_rslt <- check_dcon(dcon_tbl = ndq::dcon_input_omop,
                           compute_level = 'patient',
                           omop_or_pcornet = 'omop',
                           check_string = 'dcon')
}

## To execute the check at the visit level:
\dontrun{
my_dcon_rslt <- check_dcon(dcon_tbl = ndq::dcon_input_omop,
                           compute_level = 'visit',
                           omop_or_pcornet = 'omop',
                           check_string = 'dcon')
}

}
