% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_plausibility.R
\name{check_dp}
\alias{check_dp}
\title{Date Plausibility}
\usage{
check_dp(
  dp_tbl,
  omop_or_pcornet = "omop",
  visit_tbl = cdm_tbl("visit_occurrence"),
  dob_tbl = cdm_tbl("person"),
  check_string = "dp"
)
}
\arguments{
\item{dp_tbl}{table with information about the fact types that should
be evaluated in comparison to visit & birth dates

\if{html}{\out{<div class="sourceCode">}}\preformatted{          see ?ndq::dp_input_omop for more details
}\if{html}{\out{</div>}}}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{visit_tbl}{the CDM table with visit identifiers and dates; defaults to \code{cdm_tbl(visit_occurrence)}}

\item{dob_tbl}{the CDM table with patient identifiers and birth dates; defaults to \code{cdm_tbl(person)}}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{dp}}
}
\value{
a table with the count & proportion of rows with an implausible date value,
meaning that it fell before the associated visit start date, after the associated
visit end date, or before the patient birth date
}
\description{
This function will iterate through the provided input table to identify
the proportion of rows in each fact type that have an implausible date.
Implausibility is defined as a date that falls before the associated visit
start date, after the associated visit end date, or before the patient's
birth date.
}
