% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_concepts_present.R
\name{check_ecp}
\alias{check_ecp}
\title{Expected Concepts Present}
\usage{
check_ecp(ecp_tbl, omop_or_pcornet = "omop", check_string = "ecp")
}
\arguments{
\item{ecp_tbl}{\emph{tabular input} || \strong{required}

The primary input table that contains descriptive information about the checks
to be executed by the function. It should include definitions for the concepts
that should be identified and the denominator cohort to be used for the computation.
see \code{?ecp_input_omop} or \code{?ecp_input_pcornet} for examples of the input structure}

\item{omop_or_pcornet}{\emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{check_string}{\emph{string} || defaults to \code{ecp}

An abbreviated identifier that will be used to label all output from this module}
}
\value{
This function will return a table with the total patient count, the count of
patients with a particular concept, the proportion of total patients with the
concept, plus some additional descriptive metadata
}
\description{
This function will iterate through the provided input table to identify
the count of patients who have the at least one occurrence of the concept defined in the
user-provided concept set and the proportion of patients who have the concept
based on the user-defined denominator cohort.
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::ecp_input_omop
ndq::ecp_input_pcornet

# Use this as your input to the ECP function
## To execute the check at the patient level:
\dontrun{
my_ecp_rslt <- check_ecp(ecp_tbl = ndq::ecp_input_omop,
                         omop_or_pcornet = 'omop',
                         check_string = 'ecp')
}

}
