% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts_over_time.R
\name{check_fot}
\alias{check_fot}
\title{Facts Over Time}
\usage{
check_fot(
  fot_tbl,
  omop_or_pcornet = "omop",
  compute_method = "loop",
  time_span = list("2009-01-01", today()),
  time_period = "month",
  lookback_interval = 1,
  check_string = "fot",
  visits_only = FALSE,
  distinct_visits = TRUE
)
}
\arguments{
\item{fot_tbl}{a table with information describing the fact tables that should be examined;
see \code{?fot_input_omop} or \code{?fot_input_pcornet} for details}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{compute_method}{a string input of either \code{loop} or \code{group}; controls whether the
check is executed by looping through each time period or grouping by a date field}

\item{time_span}{a list that contains the start date and end date of the time span (i.e. list(2009-01-01, 2015-01-01))}

\item{time_period}{string indicating the length of time that the time span should be divided into (i.e. months, years)}

\item{lookback_interval}{the number of time periods (defined in \code{time_period}) to look back in each interval
(i.e. 1 year, 3 months); defaults to 1}

\item{check_string}{the abbreviated name of the check; defaults to \code{fot}}

\item{visits_only}{if TRUE, counts ONLY distinct visits and not patients or rows}

\item{distinct_visits}{if TRUE, counts distinct visits as well as total counts and total patients}
}
\value{
a dataframe with one row for each time period within the specified time span for each check;
\itemize{
\item if visits_only = TRUE, will produce only counts of visits for the check + time period
\item if visits_only = FALSE and distinct_visits = TRUE, will produce counts of patients, rows, and visits for the check + time period
\item if visits_only = FALSE and distinct_visits = FALSE, will produce counts of patients and rows for the check + time period
}
}
\description{
This function will compute the number of rows, patients, and (optionally) visits associated with
the fact of interest within a specified time period. The user will supply the end points of the time span
(i.e. January 2009 - January 2024) and the time period they wish to divide it by (i.e. month, year).
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::fot_input_omop
ndq::fot_input_pcornet

# Use this as your input to the FOT function
## This check can be executed for different time period lengths, like...
### Yearly
\dontrun{
my_fot_rslt <- check_fot(fot_tbl = ndq::fot_input_omop,
                         omop_or_pcornet = 'omop',
                         compute_method = 'loop', # use 'group' for high performant DBMSs
                         time_span = list('2015-01-01', '2025-01-01'),
                         time_period = 'year',
                         lookback_interval = 1,
                         check_string = 'fot')
}

### Monthly
\dontrun{
my_fot_rslt <- check_fot(fot_tbl = ndq::fot_input_omop,
                         omop_or_pcornet = 'omop',
                         compute_method = 'loop', # use 'group' for high performant DBMSs
                         time_span = list('2015-01-01', '2025-01-01'),
                         time_period = 'month',
                         lookback_interval = 1,
                         check_string = 'fot')
}

### Quarterly
\dontrun{
my_fot_rslt <- check_fot(fot_tbl = ndq::fot_input_omop,
                         omop_or_pcornet = 'omop',
                         compute_method = 'loop', # use 'group' for high performant DBMSs
                         time_span = list('2015-01-01', '2025-01-01'),
                         time_period = 'month',
                         lookback_interval = 3,
                         check_string = 'fot')
}

}
