% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts_over_time.R
\name{check_fot_group}
\alias{check_fot_group}
\title{Facts Over Time (Optimized for Trino)}
\usage{
check_fot_group(
  fot_tbl,
  time_frame = time_span,
  lookback_weeks = 0,
  lookback_months = 1,
  check_string = "fot",
  visits_only = TRUE,
  distinct_visits = TRUE
)
}
\arguments{
\item{fot_tbl}{a table with information describing the fact tables that should be examined}

\item{time_frame}{a list that contains at least end date of the first and last month of the desired time span}

\item{lookback_weeks}{if lookback is in weeks instead of months, this should be set to a non-zero integer.
Defaults to 0, for lookback to be in months.}

\item{lookback_months}{the number of months to look back; defaults to 1}

\item{check_string}{the abbreviated name of the check; defaults to \code{fot}}

\item{visits_only}{if TRUE, counts ONLY distinct visits and not patients}

\item{distinct_visits}{if TRUE, counts distinct visits as well as total counts and total patients}
}
\value{
table with the following rows (if \code{distinct_visits} = \code{TRUE}:
\code{month_end} | \code{check_name} | \code{database_version} | \code{site} | \code{time_desc} | \code{row_cts} | \code{row_pts} | \code{row_visits}

** if \code{time_tbls} contains fields that are grouped, the output will contain the grouped variables
}
\description{
Facts Over Time (Optimized for Trino)
}
