% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_visit_id.R
\name{check_mf_visitid}
\alias{check_mf_visitid}
\title{Missing Field: Visit ID}
\usage{
check_mf_visitid(
  mf_tbl,
  omop_or_pcornet = "omop",
  visit_tbl = cdm_tbl("visit_occurrence"),
  check_string = "mf_visitid"
)
}
\arguments{
\item{mf_tbl}{\emph{tabular input} || \strong{required}

The primary input table that contains descriptive information about the checks
to be executed by the function. It should include definitions for the clinical fact
types with visit IDs that should be evaluated against the visit_tbl.
see \code{?mf_visitid_input_omop} or \code{?mf_visitid_input_pcornet} for examples of the input structure}

\item{omop_or_pcornet}{\emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{visit_tbl}{\emph{tabular input} || defaults to \code{cdm_tbl('visit_occurrence')}

The CDM table with the visit primary key IDs. Typically, this will
be either the OMOP \code{visit_occurrence} or PCORnet \code{encounter} table}

\item{check_string}{\emph{string} || defaults to \code{mf_visitid}

An abbreviated identifier that will be used to label all output from this module}
}
\value{
This function will return a table summarizing the total number of visits in the fact table,
the number of NULL visit IDs, and the number of visits that do not have an associated entry
in the visit table
}
\description{
This function will check to see if the \code{visit_occurrence_id}/\code{encounterid} in a given fact table
also exists in the \code{visit_occurrence}/\code{encounter} table and identify cases where these IDs
are missing entirely (NULL). There may be cases where this is expected
(for example, immunizations imported from an external registry) but generally the
\code{visit_occurrence_id}/\code{encounterid} should be populated and exist as a primary key in the
\code{visit_occurrence}/\code{encounter} table.
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::mf_input_omop
ndq::mf_input_pcornet

# Use this as your input to the MF: Visit ID function
\dontrun{
my_mf_rslt <- check_mf(mf_tbl = ndq::mf_input_omop,
                       omop_or_pcornet = 'omop',
                       visit_tbl = cdm_tbl('visit_occurrence'), ## table with visit ID primary keys
                       check_string = 'mf_visitid')
}

}
