% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitative_variable_distribution.R
\name{check_qvd}
\alias{check_qvd}
\title{Quantitative Variable Distribution}
\usage{
check_qvd(
  qvd_tbl,
  qvd_cohort,
  compute_trimmed = TRUE,
  omop_or_pcornet = "omop",
  check_string = "qvd"
)
}
\arguments{
\item{qvd_tbl}{\emph{tabular input} || \strong{required}

The primary input table that contains descriptive information about the variables
to be examined by the function. It should define each variable and point the
system to the appropriate column with quantitative information.

see \code{?qvd_input_omop} or \code{?qvd_input_pcornet} for examples of the input structure}

\item{qvd_cohort}{\emph{tabular input} || \strong{required}

A table defining the specific subset of patients for which variable information
should be summarised. A more specific cohort of patients is recommended to
make it easier to assess whether certain values are plausible
(i.e. height or lab results)}

\item{compute_trimmed}{\emph{boolean} || defaults to \code{TRUE}

A boolean indicating whether the variable of interest should also be evaluated
using a trimmed distribution, which removes the top and bottom 2.5\% to lessen
the impact of the most extreme outliers on the summary statistic. These results
will be treated as an additional variable with its own row of results, rather than
adding new columns to the output.}

\item{omop_or_pcornet}{\emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{check_string}{\emph{string} || defaults to \code{dp}

An abbreviated identifier that will be used to label all output from this module}
}
\value{
This function will return a dataframe with summary statistics that describe
the distribution for each user-defined variable. This will include mean,
median, standard deviation, Q1, Q3, min, and max. Two measures of skew
will also be returned: skewness and Pearson's skew.
}
\description{
This function will, for a given cohort, compute summary statistics that
describe the numerical distribution of a given quantitative variable. If
compute_trimmed is set to TRUE, the function will additionally compute
a summary for a distribution that has the top and bottom 2.5\% removed,
to help visualize the results without the most extreme outliers.
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::qvd_input_omop
ndq::qvd_input_pcornet

# Use this as your input to the QVD function
\dontrun{
my_qvd_rslt <- check_qvd(qvd_tbl = ndq::dp_input_omop,
                         qvd_cohort = my_cohort,
                         compute_trimmed = TRUE,
                         omop_or_pcornet = 'omop',
                         check_string = 'qvd')
}

}
