% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmapped_concepts.R
\name{check_uc}
\alias{check_uc}
\title{Unmapped Concepts}
\usage{
check_uc(
  uc_tbl,
  omop_or_pcornet = "omop",
  by_year = FALSE,
  produce_mapped_list = TRUE,
  unmapped_values = c(44814650L, 0L, 44814653L, 44814649L),
  check_string = "uc"
)
}
\arguments{
\item{uc_tbl}{\emph{tabular input} || \strong{required}

The primary input table that contains descriptive information about the checks
to be executed by the function. It should include definitions for the fields that
should be evaluated for missingness and other relevant metadata.
see \code{?uc_input_omop} or \code{?uc_input_pcornet} for examples of the input structure}

\item{omop_or_pcornet}{\emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{by_year}{\emph{boolean} || defaults to \code{FALSE}

A boolean indicating whether the analysis should be conducted longitudinally by year

\strong{Note:} the mapped list functionality is NOT available for the longitudinal analysis}

\item{produce_mapped_list}{\emph{boolean} || defaults to \code{TRUE}

A boolean indicating whether a list of source values associated with unmapped concepts
should be produced in addition to the primary output. This table will be iteratively output
to the database backend of choice and will NOT be stored locally.
Only source values with > 10 occurrences are included.}

\item{unmapped_values}{\emph{string / vector} || defaults to 44814650, 0, 44814653, & 44814649

A string or vector listing the concept(s) that indicate an unmapped value. The function
will check for NULL values by default, so that does not need to be defined here.}

\item{check_string}{\emph{string} || defaults to \code{uc}

An abbreviated identifier that will be used to label all output from this module}
}
\value{
If \code{by_year} is \code{FALSE}, this function will produce a dataframe with the total row count,
the unmapped row count, the proportion of unmapped values, and some additional descriptive
metadata for each user-defined check

If \code{by_year} is \code{TRUE}, this function will produce a dataframe with the total row count,
the unmapped row count, the proportion of unmapped values, and some additional descriptive
metadata for each user-defined check stratified by each year present in the fact table

If \code{produce_mapped_list} is \code{TRUE}, then a table with name \code{uc_grpd} that includes
the source values (with > 10 appearances) and counts of those values associated
with unmapped concepts will be produced in addition to the primary tabular output
}
\description{
This function will evaluate the count and proportion of unmapped concepts associated
with the fact type of interest. If \code{produce_mapped_list} is set to TRUE, a summary of
the source values associated with unmapped concepts will also be produced to help
identify areas where mappings could potentially be improved. This analysis can be also be
executed longitudinally by year.
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::uc_input_omop
ndq::uc_input_pcornet

# Use this as your input to the UC function
## Overall
\dontrun{
my_uc_rslt <- check_uc(uc_tbl = ndq::uc_input_omop,
                       by_year = FALSE,
                       produce_mapped_list = TRUE,
                       unmapped_values = c(44814650L,0L,
                                           44814653L, 44814649L),
                       omop_or_pcornet = 'omop',
                       check_string = 'uc')
}

## By Year
\dontrun{
my_uc_rslt <- check_uc(uc_tbl = ndq::uc_input_omop,
                       by_year = TRUE,
                       produce_mapped_list = FALSE,
                       unmapped_values = c(44814650L,0L,
                                           44814653L, 44814649L),
                       omop_or_pcornet = 'omop',
                       check_string = 'uc')
}

}
