% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmapped_concepts.R
\name{check_uc_by_year}
\alias{check_uc_by_year}
\title{Unmapped Concepts by Year}
\usage{
check_uc_by_year(
  uc_tbl,
  unmapped_values = c(44814650L, 0L, 44814653L, 44814649L),
  check_string = "uc"
)
}
\arguments{
\item{uc_tbl}{dataframe with metadata describing the tables/columns for which
unmapped concepts should be identified}

\item{unmapped_values}{concepts / other values that indicate an unmapped value}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{uc}}
}
\value{
a dataframe with the following columns:
\itemize{
\item site
\item year_date
\item total_unmapped_row_ct
\item total_row_ct
\item unmapped_description
\item check_name
\item database_version
}
}
\description{
Function that produces output that contains unmapped values by year
}
