% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_vc_vs.R, R/vocabulary_conformance.R
\name{check_vc}
\alias{check_vc}
\title{Vocabulary Conformance}
\usage{
check_vc(
  vc_tbl,
  omop_or_pcornet = "omop",
  check_string = "vc",
  null_values = c(44814650L, 0L, 44814653L, 44814649L)
)

check_vc(
  vc_tbl,
  omop_or_pcornet = "omop",
  check_string = "vc",
  null_values = c(44814650L, 0L, 44814653L, 44814649L)
)
}
\arguments{
\item{vc_tbl}{a table with the table, field, and vocabulary information for each
check}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{vc}}

\item{null_values}{a vector of NULL values (or other values that would not belong to accepted vocabularies
but are broadly accepted) that should be excluded when identifying non-valueset concepts}

\item{vocabvals}{a list that contains the following values:
\itemize{
\item name of list element: the check name identifier
\enumerate{
\item the acceptable vocabulary for the field of interest, as they appear in the OMOP concept table
\item the field in which vocabulary violations should be identified
\item the name of the CDM table where the field is located
}
}}

\item{string_tbl_name}{the table name prefixed to the output}
}
\value{
a dataframe with summary information about each vocabulary that appears in the field,
with violations marked in a T/F field

\if{html}{\out{<div class="sourceCode">}}\preformatted{    vocabularies associated with "null" concepts are ignored, so proportions may not add
    up to 1 as a result
}\if{html}{\out{</div>}}

a dataframe with summary information about each vocabulary that appears in the field,
with violations marked in a T/F field

\if{html}{\out{<div class="sourceCode">}}\preformatted{    vocabularies associated with the indicated null_values are ignored, so proportions may not add
    up to 1 as a result
}\if{html}{\out{</div>}}
}
\description{
Vocabulary Conformance

Vocabulary Conformance
}
