% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_vc_vs_output}
\alias{create_vc_vs_output}
\title{Additional processing for VS & VC checks}
\usage{
create_vc_vs_output(tbl_list, check_string = "vc")
}
\arguments{
\item{tbl_list}{a list that contains all the vc or vs violations}

\item{check_string}{a string that contains the table name for the check output}
}
\value{
a pivoted version of the input table with dummy rows added for checks
that did not return any violations
}
\description{
Additional processing for VS & VC checks
}
\keyword{internal}
