% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dc_input_pcornet}
\alias{dc_input_pcornet}
\title{Data Cycle Changes Input File (PCORnet)}
\format{
\subsection{\code{dc_input_pcornet}}{

A dataframe or CSV file with 8 columns
\describe{
\item{schema_current}{The schema where the data reflecting the current data model version is kept. Use \code{cdm} to use the pre-configured \code{cdm_schema}, \code{result} to use the preconfigured \code{results_schema}, or input the exact name of the schema.}
\item{table_current}{The name of the CDM or pre-computed results table to be used to retrieve the counts associated with the current data model}
\item{schema_prev}{\strong{ONLY REQUIRED IF PREV_CT_SRC = CDM}; The schema where the data reflecting the previous data model version is kept. Use \code{cdm} to use the pre-configured \code{cdm_schema}, \code{result} to use the preconfigured \code{results_schema}, or input the exact name of the schema.}
\item{table_prev}{\strong{ONLY REQUIRED IF PREV_CT_SRC = CDM}; The name of the CDM or pre-computed results table to be used to retrieve the counts associated with the previous data model}
\item{filter_logic}{\strong{OPTIONAL} The logic that should be applied to BOTH table_current and table_prev in order to tailor the tables to the desired check assessment}
\item{check_domain}{A string indicating the domain of the check (ex: medadmin_drugs, prescription_drugs)}
\item{check_id}{A short string "code" used to identify the specific check (ex: dr_admin, dr_rx)}
\item{check_description}{A longer description of the check (ex: full med_admin table, full prescribing table)}
}
}
}
\usage{
dc_input_pcornet
}
\description{
A sample version of an PCORnet compliant input file for
the Data Cycle Changes data quality module. This exact
file is also included as a CSV in the package if the user wishes
to use it, or the structure can be copied to produce a custom
list of checks.
}
\keyword{datasets}
