% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dcon_input_omop}
\alias{dcon_input_omop}
\title{Domain Concordance Input File (OMOP)}
\format{
\subsection{\code{dcon_input_omop}}{

A dataframe or CSV file with 10 columns
\describe{
\item{cohort_id}{A short string describing the event in that row of the input file}
\item{check_id}{A short string "code" used to identify the specific check (ex: cancer_oncologists, htn_dx_antihtn_rx). There should always be \strong{TWO rows with the SAME check_id}, with each row representing one of the two input cohorts.}
\item{check_description}{A longer description of the check (ex: cancer diagnoses and oncology specialist visits, hypertension diagnoses and antihypertensive prescriptions)}
\item{schema}{The schema where the data is kept. Use \code{cdm} to use the pre-configured \code{cdm_schema}, \code{result} to use the preconfigured \code{results_schema}, or input the exact name of the schema.}
\item{table}{The name of the CDM or pre-computed results table where the relevant data is kept}
\item{date_field}{The date field to be used when time_between_events is not null to determine the length of time between each event occurrence}
\item{concept_field}{The *_concept_id field with the concepts that make up the valueset (ex: condition_concept_id, drug_concept_id)}
\item{conceptset_name}{\strong{OPTIONAL} The string name of the concept set that will identify the concepts of interest, as it appears in the predefined file_subdirectory}
\item{filter_logic}{\strong{OPTIONAL} The logic that should be applied to the provided table in order to tailor the tables to the desired check assessment (ex: if you only want to assess prescription drugs)}
\item{time_between_events}{\strong{OPTIONAL} An integer expressing the \emph{maximum} number of days that should fall between the two events. Patients with events that occur too far apart will not be included in the combined cohort.}
}
}
}
\usage{
dcon_input_omop
}
\description{
A sample version of an OMOP compliant input file for
the Domain Concordance data quality module. This exact
file is also included as a CSV in the package if the user wishes
to use it, or the structure can be copied to produce a custom
list of checks.
}
\details{
If you choose to use this sample file, please be sure to
set your \code{file_subdirectory}
to \code{system.file("sample_conceptsets", package = "ndq")} so
the functions know where to access the associated concept sets. You
also have the option to download the sample concept sets to a local
directory and point to that location.
}
\keyword{datasets}
