% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecp_input_omop}
\alias{ecp_input_omop}
\title{Expected Concepts Present Input File (OMOP)}
\format{
\subsection{\code{ecp_input_omop}}{

A dataframe or CSV file with 9 columns
\describe{
\item{check_id}{A short string "code" used to identify the specific check (ex: de, de_rx)}
\item{cohort_definition}{The definition of the specific cohort used to identify patients with the concept of interest}
\item{cohort_schema}{The schema where the cohort_table is kept. Use \code{cdm} to use the pre-configured \code{cdm_schema}, \code{result} to use the preconfigured \code{results_schema}, or input the exact name of the schema.}
\item{cohort_table}{The name of the CDM or pre-computed results table where the cohort has been stored}
\item{schema}{The schema where the data is kept. Use \code{cdm} to use the pre-configured \code{cdm_schema}, \code{result} to use the preconfigured \code{results_schema}, or input the exact name of the schema.}
\item{table}{The name of the CDM or pre-computed results table where the relevant data is kept}
\item{concept_field}{The *_concept_id field with the concepts that make up the valueset (ex: drug_concept_id, drug_source_concept_id)}
\item{conceptset_name}{The string name of the concept set that will identify the concepts of interest, as it appears in the predefined file_subdirectory}
\item{filter_logic}{\strong{OPTIONAL} The logic that should be applied to the provided table in order to tailor the tables to the desired check assessment (ex: if you only want to identify prescriptions for antihypertensives)}
}
}
}
\usage{
ecp_input_omop
}
\description{
A sample version of an OMOP compliant input file for
the Unmapped Concepts data quality module. This exact
file is also included as a CSV in the package if the user wishes
to use it, or the structure can be copied to produce a custom
list of checks.
}
\details{
If you choose to use this sample file, please be sure to
set your \code{file_subdirectory}
to \code{system.file("sample_conceptsets", package = "ndq")} so
the functions know where to access the associated concept sets. You
also have the option to download the sample concept sets to a local
directory and point to that location.
}
\keyword{datasets}
