% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{find_concept_names}
\alias{find_concept_names}
\title{Find concept names for existing concept_ids}
\usage{
find_concept_names(
  fact_tbl,
  omop_or_pcornet,
  fact_concept_id,
  concept_field,
  concept_tbl = vocabulary_tbl("concept")
)
}
\arguments{
\item{fact_tbl}{the fact table associated with the field}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{fact_concept_id}{the concept field in the fact table (i.e. condition_concept_id or dx)}

\item{concept_field}{the field in the concept table to be used in the analysis; will typically be
concept_name, but concept_class_id is used for RxNorm class identifiers}

\item{concept_tbl}{defaults to \code{vocabulary_tbl('concept')}}
}
\value{
the provided fact_tbl with additional information from the specified concept_field
}
\description{
Find concept names for existing concept_ids
}
