% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{output_tbl_append}
\alias{output_tbl_append}
\title{output table to database if it does not exist, or
append it to an existing table with the same name if it does}
\usage{
output_tbl_append(
  data,
  name = NA,
  local = FALSE,
  file = base::ifelse(config("results_target") == "file", TRUE, FALSE),
  db = if (!file) config("results_target") else NA,
  results_tag = TRUE,
  ...
)
}
\arguments{
\item{data}{the data to output}

\item{name}{the name of the table to output

Parameters are the same as \code{output_tbl}}
}
\value{
The table as it exists on the databse, with the new data
appended, if the table already existts.
}
\description{
output table to database if it does not exist, or
append it to an existing table with the same name if it does
}
