% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{process_bmc}
\alias{process_bmc}
\title{Best Mapped Concepts -- Processing}
\usage{
process_bmc(
  bmc_results,
  bmc_concepts_labelled,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{bmc_results}{the bmc_counts table output by check_bmc}

\item{bmc_concepts_labelled}{the bmc_concepts table output by check_bmc, with an additional
column called "include" added with "not best" or unideal concepts
marked with a 0}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
a dataframe summarizing the proportion of best vs not best concepts for a given check, indicated by
the appropriate value in the "include" column
}
\description{
Best Mapped Concepts -- Processing
}
