% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_fact_documentation.R
\name{process_cfd}
\alias{process_cfd}
\title{Clinical Fact Documentation -- Processing}
\usage{
process_cfd(cfd_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{cfd_results}{table output by check_cfd}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
cfd_output tbl with additional total counts
}
\description{
Intakes the output of check_cfd in order to apply additional processing. This
includes computing overall counts/proportions across all sites included in the input and tidying
some of the descriptive metadata.
}
