% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_concordance.R
\name{process_dcon}
\alias{process_dcon}
\title{Domain Concordance -- Processing}
\usage{
process_dcon(dcon_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{dcon_results}{table output by check_dcon}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
a dataframe with one row for cohort overlap computation type listed in the description, with
the associated raw count (of patients or visits) and the associated proportion
}
\description{
Intakes the output of check_dcon in order to apply additional processing. This
includes computing the following cohort overlap counts and proportions:
\itemize{
\item cohort_1_only: overall, patients in just cohort 1
\item cohort_2_only: overall, patients in just cohort 2
\item combined: overall, patients in both 1 and 2
\item cohort_1_denom: patients in cohort 1 not cohort 2
\item cohort_2_denom: patients in cohort 2 not cohort 1
\item cohort_1_in_2: patients in cohort 2 who are also in 1 (use cohort 2 as denominator)
\item cohort_2_in_1: patients in cohort 1 who are also in 2 (use cohort 1 as denominator)
}
}
\details{
Note that cohort_1_in_2 and cohort_2_in_1 will have the same raw count, but different
proportions since the denominator is different
}
\examples{
# This function should be run after check_dcon has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_dcon_process <- process_dcon(dcon_results = my_dcon_rslts,
                                rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_dcon_process <- process_dcon(dcon_results = 'my_dcon_rslts',
                                rslt_source = 'csv',
                                csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_dcon_process <- process_dcon(dcon_results = 'my_dcon_rslts',
                                rslt_source = 'remote')
}

}
