% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts_over_time.R
\name{process_fot}
\alias{process_fot}
\title{Facts Over Time -- Processing}
\usage{
process_fot(
  fot_results,
  target_col = "row_cts",
  add_ratios = FALSE,
  ratio_mult = 10000,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{fot_results}{the table output by check_fot}

\item{target_col}{the numerical column in fot_results that should be used to
compute the heuristic; options are row_cts, row_pts, or row_visits}

\item{add_ratios}{boolean to indicate whether ratios / rates should be computed
if TRUE, the fot_results table should have an additional \code{total_pt}
column with counts for the denominator of the user's choosing

\if{html}{\out{<div class="sourceCode">}}\preformatted{              for example, the count of patients in that time period that
              had a visit
}\if{html}{\out{</div>}}}

\item{ratio_mult}{if add_ratios = TRUE, the numerical multiplier that should
be used to compute the rate}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
a list with up to three dataframes:
\code{fot_heuristic}: the summary heuristic (month / ((month-1)\emph{.25 +
(month+1)}.25 +
(month-12)*.5))
\code{fot_heuristic_summary}: summary values (mean, med, sd, q1, q3) based on
the heuristic
and if add_ratios = TRUE
\code{fot_ratios}: the fot_results table with an additional column with the
incidence ratio applied (as row_ratio)
}
\description{
Intakes the output of check_fot in order to apply additional processing. This
includes applying a heuristic meant to compare the fact count in a given time period
to other time periods around it. For a montly computation, for example, the heuristic
would be \code{month / ((month-1 * .25) + (month+1 * .25) + (month-12 * .5))} which is
the value for a given month divided by the weighted average of the value in the previous
month, the next month, and the same month in the previous year. The function can also
optionally compute a ratio given a \code{total_pt} column (added by the user) to be used as a
denominator and a \code{ratio_mult} value to be used as a multiplier.
}
\examples{
# This function should be run after check_fot has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_fot_process <- process_fot(fot_results = my_fot_rslts,
                              target_col = 'row_cts',
                              rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_fot_process <- process_fot(fot_results = 'my_fot_rslts',
                              target_col = 'row_cts',
                              rslt_source = 'csv',
                              csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_fot_process <- process_fot(fot_results = 'my_fot_rslts',
                              target_col = 'row_cts',
                              rslt_source = 'remote')
}

# You can also optionally compute patient incidence ratios. This computation will use
# the patient count, regardless of what is specified in the `target_col` argument.
# A column with the total patient count for that time period will need to be added to
# the results output. It can reflect any patient cohort, so long as the column is called
# `total_pt`. We recommend computing the denominator as part of the initial check_fot execution
# and extracting the counts from the resulting output.

\dontrun{
my_fot_process <- process_fot(fot_results = 'my_fot_rslts',
                              target_col = 'row_cts',
                              add_ratios = TRUE,
                              ratio_mult = 10000,
                              rslt_source = 'remote')
}

}
