% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_visit_id.R
\name{process_mf_visitid}
\alias{process_mf_visitid}
\title{Missing Field: Visit ID -- Processing}
\usage{
process_mf_visitid(
  mf_visitid_results,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{mf_visitid_results}{output of check_mf_visitid}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
mf_visitid tbl with additional domain, total_ct, and proportion
column
}
\description{
function to add proportions to mf results
}
