% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_facts.R
\name{process_pf}
\alias{process_pf}
\title{Patient Facts -- Processing}
\usage{
process_pf(pf_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{pf_results}{table output by check_pf}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
pf_output tbl with additional total counts
}
\description{
Intakes the output of check_pf in order to apply additional processing. This
includes computing overall counts/proportions across all sites included in the input and tidying
some of the descriptive metadata.
}
