% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmapped_concepts.R
\name{process_uc}
\alias{process_uc}
\title{Unmapped Concepts -- Processing}
\usage{
process_uc(uc_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{uc_results}{the output of check_uc}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
table with additional columns/etc needed for pp output
}
\description{
Function to add post-processed columns to the unmapped concepts dqa_library output
and to add a \code{total} row with for each of the check applications
for the overall number and proportion of unmapped rows
}
