% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_conformance.R
\name{process_vc}
\alias{process_vc}
\title{Vocabulary Conformance -- Processing}
\usage{
process_vc(vc_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{vc_results}{\emph{tabular input} || \strong{required}

The tabular output of \code{check_vc}. This table should include results for all
institutions that should be included in the computation of overall / "network level"
statistics.}

\item{rslt_source}{\emph{string} || defaults to \code{remote}

A string that identifies the location of the \code{vc_results} table.
Acceptable values are
\itemize{
\item \code{local} - table is stored as a dataframe in the local R environment
\item \code{csv} - table is stored as a CSV file
\item \code{remote} - table is stored on a remote database
}}

\item{csv_rslt_path}{\emph{string} || defaults to \code{NULL}

If \code{rslt_source} has been set to \code{csv}, this parameter should indicate the path to
the result file(s). Otherwise, this parameter can be left as \code{NULL}}
}
\value{
This function will return a list that contains two dataframes:
-\code{vc_processed}: The \code{vc_results} table with additional rows reflecting the overall / "network level"
counts and a computed proportion of violating vocabularies
-\code{vc_violations}: A table listing all of the vocabulary violations
}
\description{
Intakes the output of \code{check_vc} in order to apply additional processing. This
includes computing row and patient proportions and computing overall totals
across all sites included in the input.
}
\examples{
# This function should be run after check_vc has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_vc_process <- process_vc(vc_results = my_vc_rslts,
                            rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_vc_process <- process_vc(vc_results = 'my_vc_rslts',
                            rslt_source = 'csv',
                            csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_vc_process <- process_vc(vc_results = 'my_vc_rslts',
                            rslt_source = 'remote')
}

}
