% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueset_conformance.R
\name{process_vs}
\alias{process_vs}
\title{Valueset Conformance -- Processing}
\usage{
process_vs(vs_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{vs_results}{the output of check_vs}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
a list that contains two dataframes:
table with:
site
table_application
measurement_column
vocabulary_id
check_type
check_name
total_denom_ct
accepted_value
tot_ct: sum of total rows for given value in vs
tot_prop: sum of proportion of all values for that vs

\if{html}{\out{<div class="sourceCode">}}\preformatted{     and another table with a similar structure but only containing the valueset violations
}\if{html}{\out{</div>}}
}
\description{
Valueset Conformance -- Processing
}
