% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_n_summary.R
\name{summarize_large_n}
\alias{summarize_large_n}
\title{Add summary columns for a large N of sites}
\usage{
summarize_large_n(
  dq_output,
  num_col,
  grp_vars,
  check_string,
  time = FALSE,
  shape
)
}
\arguments{
\item{dq_output}{the output table from processing to which the columns
should be added}

\item{num_col}{the numeric column to be used to compute the statistics}

\item{grp_vars}{grouping variables for the computation}

\item{check_string}{string indicating the check type}

\item{time}{a logical indicating whether the check is time dependent}

\item{shape}{'long' or 'wide' indicating whether summary statistics should be separate columns (wide)
or assigned to separate rows, with the name of the summary statistic in the site column (long)}
}
\value{
a dataframe with summary statistics for the numeric column
provided. contains all columns from \code{dq_output} plus:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if time = `TRUE`, the mean and median are returned.

    if time = `FALSE`, the max, min, q1, q3, mean, and median are returned

    These values are treated as "sites" to assist with Shiny implementation
}\if{html}{\out{</div>}}
}
\description{
Add summary columns for a large N of sites
}
