% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vc_input_omop}
\alias{vc_input_omop}
\title{Vocabulary Conformance Input File (OMOP)}
\format{
\subsection{\code{vc_input_omop}}{

A dataframe or CSV file with 6 columns
\describe{
\item{check_id}{A short string "code" used to identify the specific check (ex: de, de_rx)}
\item{schema}{The schema where the data is kept. Use \code{cdm} to use the pre-configured \code{cdm_schema}, \code{result} to use the preconfigured \code{results_schema}, or input the exact name of the schema.}
\item{table}{The name of the CDM or pre-computed results table where the relevant data is kept}
\item{acceptable_vocabularies}{A comma-separated string of acceptable vocabulary values (ex: "RxNorm, RxNorm Extension")}
\item{concept_field}{The *_concept_id field with the concepts that should be evaluated (ex: drug_concept_id, drug_source_concept_id)}
\item{filter_logic}{\strong{OPTIONAL} The logic that should be applied to the provided table in order to tailor the tables to the desired check assessment (ex: if prescription drugs had different vocabulary requirements than inpatient administrations)}
}
}
}
\usage{
vc_input_omop
}
\description{
A sample version of an OMOP compliant input file for
the Vocabulary Conformance data quality module. This exact
file is also included as a CSV in the package if the user wishes
to use it, or the structure can be copied to produce a custom
list of checks.
}
\keyword{datasets}
