% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs_input_pcornet}
\alias{vs_input_pcornet}
\title{Valueset Conformance Input File (PCORnet)}
\format{
\subsection{\code{vs_input_pcornet}}{

A dataframe or CSV file with 7 columns
\describe{
\item{check_id}{A short string "code" used to identify the specific check (ex: de, de_rx)}
\item{check_description}{A longer description of the check}
\item{schema}{The schema where the data is kept. Use \code{cdm} to use the pre-configured \code{cdm_schema}, \code{result} to use the preconfigured \code{results_schema}, or input the exact name of the schema.}
\item{table}{The name of the CDM or pre-computed results table where the relevant data is kept}
\item{valueset_name}{The name of the CSV file (stored in the pre-defined file subdirectory) defining the acceptable values for the valueset}
\item{concept_field}{The *_concept_id field with the concepts that make up the valueset (ex: race, hispanic)}
\item{vocabulary_field}{Typically only relevant for diagnoses and procedures, the field where the code vocabulary is kept (i.e. dx_type, px_type)}
\item{filter_logic}{\strong{OPTIONAL} The logic that should be applied to the provided table in order to tailor the tables to the desired check assessment (ex: if you only want to evaluate the valueset for billed conditions)}
}
}
}
\usage{
vs_input_pcornet
}
\description{
A sample version of an PCORnet compliant input file for
the Valueset Conformance data quality module. This exact
file is also included as a CSV in the package if the user wishes
to use it, or the structure can be copied to produce a custom
list of checks.
}
\keyword{datasets}
