% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_meta}
\alias{add_meta}
\title{add check name, db version, and site name to a given table}
\usage{
add_meta(
  tbl_meta,
  check_lib,
  version = config("current_version"),
  site_nm = config("site")
)
}
\arguments{
\item{tbl_meta}{the table to add meta information to}

\item{check_lib}{the name of the check}

\item{version}{the version of the database; defaults to
\code{config('current_version')};}

\item{site_nm}{the name of the site; defaults to
\code{config('site')}}
}
\description{
add check name, db version, and site name to a given table
}
