% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_anonymization.R
\name{attach_anon_id}
\alias{attach_anon_id}
\title{Add an anonymized site identifier to all tables}
\usage{
attach_anon_id(all_sites_tbl, tbls_to_anon)
}
\arguments{
\item{all_sites_tbl}{\emph{tabular input} || \strong{required}

A reference table that contains a column called \code{site} with all of
the sites that should be masked in each of the provided result tables}

\item{tbls_to_anon}{\emph{named list} || \strong{required}

A \strong{NAMED} list of all of the tables that should be anonymized}
}
\value{
This function will return each of the original tables in \code{tbls_to_anon} with
all the original columns plus a column called \code{site_anon} with a masked identifier
that is consistent across all of the tables in \code{tbls_to_anon}
}
\description{
This function will iterate through the provided list of result tables
to apply consistent anonymization of site names across all output.
}
