% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{check_bmc}
\alias{check_bmc}
\title{Best Mapped Concepts}
\usage{
check_bmc(
  bmc_tbl,
  omop_or_pcornet = "omop",
  concept_tbl = vocabulary_tbl("concept"),
  check_string = "bmc"
)
}
\arguments{
\item{bmc_tbl}{table with information regarding the fields that should be
evaluated to determine if they only include "best" concepts}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{concept_tbl}{the vocabulary table with concept information; defaults to 'concept'}

\item{check_string}{string that contains a description of the table}
}
\value{
a list of two dataframes:
one dataframe that contains each concept present in the field
and the associated row and patient counts/proportions

\if{html}{\out{<div class="sourceCode">}}\preformatted{    and another dataframe with just the concepts to be used to identify
    "best" vs "not best" concepts in the processing step
}\if{html}{\out{</div>}}
}
\description{
compute numbers proportion of rows and patients at
each level of rxnorm drug level
}
