% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{check_bmc}
\alias{check_bmc}
\title{Best Mapped Concepts}
\usage{
check_bmc(
  bmc_tbl,
  omop_or_pcornet = "omop",
  concept_tbl = NULL,
  check_string = "bmc"
)
}
\arguments{
\item{bmc_tbl}{\emph{tabular input} || \emph{required}

The primary input table that contains descriptive information about the checks
to be executed by the function. It should include definitions the fields that should be
evaluated to determine if they only include "best" concepts.
see \code{?bmc_input_omop} or \code{?bmc_input_pcornet} for examples of the input structure}

\item{omop_or_pcornet}{\emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{concept_tbl}{\emph{tabular input} || defaults to \code{NULL}

An optional parameter used to define a vocabulary table with concept definitions
(for example, the OHDSI concept table). If left NULL, the concepts as they exist
in the fact table will be returned to the user.}

\item{check_string}{\emph{string} || defaults to \code{bmc}

An abbreviated identifier that will be used to label all output from this module}
}
\value{
This function will return a list of two dataframes:
\itemize{
\item \code{bmc_counts}: A table with one row for each concept present in each user-defined field
and the associated row and patient counts/proportions
\item \code{bmc_concepts}: A table with just the concepts from \code{bmc_counts}. This output is
should be labelled with "best" (1) vs "not best" (0) indicators in a column
called \code{include} for use in the processing step
}
}
\description{
This function will identify the existing concepts within the specified field
so the user can assess which of these concepts are acceptable ("best") or should
not be used in that field ("not best")
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::bmc_input_omop
ndq::bmc_input_pcornet

# Use this as your input to the BMC function
\dontrun{
my_bmc_rslt <- check_bmc(bmc_tbl = ndq::bmc_input_omop,
                         omop_or_pcornet = 'omop',
                         concept_tbl = vocabulary_tbl("concept"), ## points to OHDSI concept table
                         check_string = 'bmc')
}


}
