% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{check_bmc}
\alias{check_bmc}
\title{Best Mapped Concepts}
\usage{
check_bmc(
  bmc_tbl,
  omop_or_pcornet = "omop",
  concept_tbl = vocabulary_tbl("concept"),
  check_string = "bmc"
)
}
\arguments{
\item{bmc_tbl}{table with information regarding the fields that should be
evaluated to determine if they only include "best" concepts;
see \code{?bmc_input_omop} or \code{?bmc_input_pcornet} for details}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{concept_tbl}{the vocabulary table with concept information; defaults to \code{vocabulary.concept}}

\item{check_string}{string that contains a description of the table}
}
\value{
a list of two dataframes:
\code{bmc_counts}: each concept present in the specified field and the
associated row and patient counts/proportions
\code{bmc_concepts}: just the concepts from bmc_counts -- to be labelled
as "best" (1) vs "not best" (0) concepts in a column
called \code{include} for use in the processing step
}
\description{
This function will identify the existing concepts within the specified field
so the user can assess which of these concepts are acceptable ("best") or should
not be used in that field ("not best")
}
