% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_concordance.R
\name{check_dcon}
\alias{check_dcon}
\title{Compute domain concordance between 2 cohorts}
\usage{
check_dcon(
  dcon_tbl,
  compute_level = "patient",
  omop_or_pcornet = "omop",
  check_string = "dcon"
)
}
\arguments{
\item{dcon_tbl}{table describing each cohort pair that should be examined;
there should be 2 rows per check, one for each cohort, with the same check_id}

\item{compute_level}{string indicating the level at which the computation should be executed
accepted values are 'patient' or 'visit'}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{check_string}{the abbreviated name of the check; defaults to \code{dcon}}
}
\value{
one dataframe with counts for the patients/visits in the first cohort,
the patients/visits in the second cohort, and the patients/visits in both

\if{html}{\out{<div class="sourceCode">}}\preformatted{    contains the columns: value, cohort, yr (set to 9999), check_type,
                          database_version, site, check_name, check_desc
}\if{html}{\out{</div>}}
}
\description{
Compute domain concordance between 2 cohorts
}
