% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_concordance.R
\name{check_dcon}
\alias{check_dcon}
\title{Domain Concordance}
\usage{
check_dcon(
  dcon_tbl,
  compute_level = "patient",
  omop_or_pcornet = "omop",
  check_string = "dcon"
)
}
\arguments{
\item{dcon_tbl}{table describing each cohort pair that should be examined;
there should be 2 rows per check, one for each cohort, with the same check_id
see \code{?dcon_input_omop} or \code{?dcon_input_pcornet} for details}

\item{compute_level}{string indicating the level at which the computation should be executed
accepted values are \code{patient} or \code{visit}}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{check_string}{the abbreviated name of the check; defaults to \code{dcon}}
}
\value{
one dataframe with counts for the patients/visits in the first cohort,
the patients/visits in the second cohort, and the patients/visits in both
}
\description{
Given the details of a pair of clinical events, this function
will determine the count of patients OR visits that meet criteria for
the first event, the second event, and both events. Users can optionally
define a time limitation for the combined cohort, so patients/visits will only
count towards that cohort when the two events occur within a specified number of days
of each other.
}
