% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_plausibility.R
\name{check_dp}
\alias{check_dp}
\title{Date Plausibility}
\usage{
check_dp(
  dp_tbl,
  omop_or_pcornet = "omop",
  visit_tbl = cdm_tbl("visit_occurrence"),
  dob_tbl = cdm_tbl("person"),
  check_string = "dp"
)
}
\arguments{
\item{dp_tbl}{\emph{tabular input} || \strong{required}

The primary input table that contains descriptive information about the checks
to be executed by the function. It should include definitions for the clinical fact
types that should be evaluated for date plausibility.
see \code{?dp_input_omop} or \code{?dp_input_pcornet} for examples of the input structure}

\item{omop_or_pcornet}{\emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{visit_tbl}{\emph{tabular input} || defaults to \code{cdm_tbl(visit_occurrence)}

A table with visit information, including identifiers that can link back to facts
and start & end dates. This argument will most likely be the \code{visit_occurrence} table for OMOP or
the \code{encounter} table for PCORnet. Custom tables with specific visit subsets
(i.e. a table with only nephrology specialist visits) can also be provided.}

\item{dob_tbl}{\emph{tabular input} || defaults to \code{cdm_tbl('person')}

A table with patient identifiers that can link back to facts and birth dates. This argument
will most likely be the \code{person} table for OMOP or the \code{demographic} table for PCORnet.}

\item{check_string}{\emph{string} || defaults to \code{dp}

An abbreviated identifier that will be used to label all output from this module}
}
\value{
This function will return a table with the count & proportion of rows with an
implausible date value, with one column for each of the 3 implausibility definitions.
These will indicate the proportion of facts falling before the associated visit start date,
after the associated visit end date, or before the patient birth date
}
\description{
This function will iterate through the provided input table to identify
the proportion of rows in each fact type that have an implausible date.
Implausibility is defined as a date that falls before the associated visit
start date, after the associated visit end date, or before the patient's
birth date.
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::dp_input_omop
ndq::dp_input_pcornet

# Use this as your input to the DP function
\dontrun{
my_dp_rslt <- check_dp(dp_tbl = ndq::dp_input_omop,
                       omop_or_pcornet = 'omop',
                       visit_tbl = cdm_tbl('visit_occurrence'), ## table with visit/encounter dates
                       dob_tbl = cdm_tbl('person'), ## table with dates of birth
                       check_string = 'dp')
}

}
