% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_concepts_present.R
\name{check_ecp}
\alias{check_ecp}
\title{Check Expected Concept Presence}
\usage{
check_ecp(ecp_tbl, omop_or_pcornet = "omop", check_string = "ecp")
}
\arguments{
\item{ecp_tbl}{table with information regarding the variables and associated fact tables
that should be evaluated}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{ecp}}
}
\value{
a table with the total patient count, the count of patients with a particular concept,
the proportion of total patients with the concept, and relevant metadata
}
\description{
This function will loop through the provided table to identify
the count of patients who have the concept identified in the list element and the proportion
of patients who have the concept based on the user-provided denominator table.
}
