% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts_over_time.R
\name{check_fot_loop}
\alias{check_fot_loop}
\title{Facts Over Time (Original Postgres Implementation)}
\usage{
check_fot_loop(
  fot_tbl,
  time_frame,
  omop_or_pcornet = "omop",
  lookback_interval = 1,
  check_string = "fot",
  visits_only = FALSE,
  distinct_visits = TRUE
)
}
\arguments{
\item{fot_tbl}{a table with information describing the fact tables that should be examined;
see \code{?fot_input_omop} or \code{?fot_input_pcornet} for details}

\item{time_frame}{a table of dates that should be iterated through to retrieve the facts for each time period;
has columns: time_start, time_end}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{lookback_interval}{the number of time periods (defined in check_fot) to look back;
defaults to 1}

\item{check_string}{the abbreviated name of the check; defaults to \code{fot}}

\item{visits_only}{if TRUE, counts ONLY distinct visits and not patients or rows}

\item{distinct_visits}{if TRUE, counts distinct visits as well as total counts and total patients}
}
\value{
a dataframe with one row for each time period within the specified time span for each check;
\itemize{
\item if visits_only = TRUE, will produce only counts of visits for the check + time period
\item if visits_only = FALSE and distinct_visits = TRUE, will produce counts of patients, rows, and visits for the check + time period
\item if visits_only = FALSE and distinct_visits = FALSE, will produce counts of patients and rows for the check + time period
}
}
\description{
Facts Over Time (Original Postgres Implementation)
}
\keyword{internal}
