% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_visit_id.R
\name{check_mf_visitid}
\alias{check_mf_visitid}
\title{Missing Field: Visit ID}
\usage{
check_mf_visitid(
  mf_tbl,
  omop_or_pcornet = "omop",
  visit_tbl = cdm_tbl("visit_occurrence"),
  check_string = "mf_visitid"
)
}
\arguments{
\item{mf_tbl}{a table with information about the tables that should be cross
checked against the visit_tbl to ensure the visit_occurrence_id / encounterid
exists as a primary key}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{visit_tbl}{the CDM visit_occurrence / encounter table that contains the visit ID primary key}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{mf_visitid}}
}
\value{
a table summarizing the total number of visits in the fact table, the number of NULL visits,
and the number of visits that cannot link back to the visit table. has columns: measure,
total_visits, missing_visits_total, missing_visits_distinct, visit_na, total_id, check_name,
database_version,site
}
\description{
This function will check to see if the visit_occurrence_id/encounterid in a given fact table
also exists in the visit_occurrence/encoubter table and identify cases where the visit_occurrence_id/encounterid
is missing entirely (NULL). There may be cases where this is expected
(for example, immunizations imported from an external registry) but generally the
visit_occurrence_id/encounterid should be populated and exist as a primary key in the
visit_occurrence/encounter table.
}
