% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmapped_concepts.R
\name{check_uc}
\alias{check_uc}
\title{Function to check for value set conformance to a defined set}
\usage{
check_uc(
  uc_tbl,
  by_year = FALSE,
  produce_mapped_list = TRUE,
  unmapped_values = c(44814650L, 0L, 44814653L, 44814649L),
  check_string = "uc"
)
}
\arguments{
\item{uc_tbl}{dataframe with metadata describing the tables/columns for which
unmapped concepts should be identified}

\item{by_year}{boolean indicating whether the analysis should be conducted longitudinally by year
or not; note that a mapped list will NOT be produced for the longitudinal analysis}

\item{produce_mapped_list}{if \code{TRUE} then will produce a table with the source values
associated with the top unmapped values; this table will
be iteratively output to the database backend of choice and will
NOT be stored locally. Only source values with > 10 occurrences
are included.}

\item{unmapped_values}{concepts / other values that indicate an unmapped value}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{uc}}
}
\value{
tbl with the following columns:
\itemize{
\item site
\item measure: from element name describing measuer
\item total_rows: the denominator
\item unmapped_rows: numerator
\item check_name: from \code{check_string}
\item database_version
\item unmapped_prop
}

if \code{produce_mapped_list} is \code{TRUE}, then the following table will be automatically output
to the database with the table name \code{st_conf_concepts_grpd}:
\itemize{
\item site
\item unmapped_description
\item database_version
\item check_name
\item src_value
\item src_value_name
\item src_value_ct
}
}
\description{
Function to check for value set conformance to a defined set
}
