% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmapped_concepts.R
\name{check_uc}
\alias{check_uc}
\title{Unmapped Concepts}
\usage{
check_uc(
  uc_tbl,
  by_year = FALSE,
  produce_mapped_list = TRUE,
  unmapped_values = c(44814650L, 0L, 44814653L, 44814649L),
  check_string = "uc"
)
}
\arguments{
\item{uc_tbl}{dataframe with metadata describing the tables/columns for which
unmapped concepts should be identified}

\item{by_year}{boolean indicating whether the analysis should be conducted longitudinally by year
or not; note that a mapped list will NOT be produced for the longitudinal analysis}

\item{produce_mapped_list}{if \code{TRUE} then will produce a table with the source values
associated with the top unmapped values; this table will
be iteratively output to the database backend of choice and will
NOT be stored locally. Only source values with > 10 occurrences
are included.}

\item{unmapped_values}{concepts / other values that indicate an unmapped value}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{uc}}
}
\value{
if \code{by_year} is \code{FALSE}: a dataframe with the total row count, the unmapped row count, the proportion
of unmapped values, and some additional descriptive metadata for each check

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if `by_year` is `TRUE`: a dataframe with the total row count, the unmapped row count, the proportion
    of unmapped values, and some additional descriptive metadata for each check stratified by each year
    present in the fact table

    if `produce_mapped_list` is `TRUE`, then a table with name `uc_grpd` that includes
    the source values (with > 10 appearances) and counts of those values associated
    with unmapped concepts
}\if{html}{\out{</div>}}
}
\description{
This function will evaluate the count and proportion of unmapped concepts associated
with the fact type of interest. If produce_mapped_list is set to TRUE, a summary of
the source values associated with unmapped concepts will also be produced to help
identify areas where mappings could potentially be improved.
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::uc_input_omop
ndq::uc_input_pcornet

# Use this as your input to the UC function
## Overall
\dontrun{
my_uc_rslt <- check_uc(uc_tbl = ndq::uc_input_omop,
                       by_year = FALSE,
                       produce_mapped_list = TRUE, # list of unmapped source values
                       unmapped_values = c(44814650L,0L,
                                           44814653L, 44814649L) # will also check for NA by default
                       omop_or_pcornet = 'omop',
                       check_string = 'uc')
}

## By Year
\dontrun{
my_uc_rslt <- check_uc(uc_tbl = ndq::uc_input_omop,
                       by_year = TRUE,
                       produce_mapped_list = FALSE, # not functionality for by year analysis
                       unmapped_values = c(44814650L,0L,
                                           44814653L, 44814649L) # will also check for NA by default
                       omop_or_pcornet = 'omop',
                       check_string = 'uc')
}

}
