% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_conformance.R
\name{check_vc}
\alias{check_vc}
\title{Vocabulary Conformance}
\usage{
check_vc(
  vc_tbl,
  omop_or_pcornet = "omop",
  check_string = "vc",
  concept_tbl = vocabulary_tbl("concept"),
  null_values = c(44814650L, 0L, 44814653L, 44814649L)
)
}
\arguments{
\item{vc_tbl}{\emph{tabular input} || \strong{required}

The primary input table that contains descriptive information about the checks
to be executed by the function. It should include definitions for database table fields and
the expected vocabularies for that field.
see \code{?vc_input_omop} or \code{?vc_input_pcornet} for examples of the input structure}

\item{omop_or_pcornet}{somop_or_pcornet \emph{string} || defaults to \code{omop}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}

\item{check_string}{\emph{string} || defaults to \code{vc}

An abbreviated identifier that will be used to label all output from this module}

\item{concept_tbl}{\emph{tabular input} || defaults to \code{vocabulary_tbl('concept')}

A vocabulary table with concept definitions (for example, the OHDSI concept table) that includes
the vocabulary to which the concept belongs. This table should at a minimum have the columns:
\code{concept_id}, \code{concept_name}, \code{vocabulary_id}}

\item{null_values}{\emph{string / vector} || defaults to 44814650, 0, 44814653, & 44814649

A string or vector listing the concept(s) that indicate a NULL value, which will be
excluding when assessing for the presence of vocabularies not accepted in the field.}
}
\value{
This function will return a table with summary information about each vocabulary that
appears in the field, with violations marked in a T/F field. These summaries are computed at
both the row and concept levels. This is to account for any cases where a large quantity of rows are
in violation of the acceptable vocabularies, but it is made up of only 1-2 distinct concepts.

Note that vocabularies associated with the indicated \code{null_values} are ignored, so proportions may
not add up to 1 as a result.
}
\description{
This function will use a provided table with concept to vocabulary mappings to
identify the vocabulary of each concept and determine how many rows comply with
the standard vocabularies expected for that field and how many rows violate these expectations.
}
\examples{
# First create input file with desired checks to be executed
# You can access examples for both OMOP & PCORnet here:
ndq::vc_input_omop
ndq::vc_input_pcornet

# Use this as your input to the vc function
\dontrun{
my_vc_rslt <- check_vc(vc_tbl = ndq::vc_input_omop,
                       omop_or_pcornet = 'omop',
                       concept_tbl = vocabulary_tbl("concept"), ## points to OHDSI concept table
                       null_values = c(44814650L,0L,
                                       44814653L,44814649L), ## ignored illegal vocabs
                       check_string = 'vc')
}


}
