% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_conformance.R
\name{check_vc}
\alias{check_vc}
\title{Vocabulary Conformance}
\usage{
check_vc(
  vc_tbl,
  omop_or_pcornet = "omop",
  check_string = "vc",
  null_values = c(44814650L, 0L, 44814653L, 44814649L)
)
}
\arguments{
\item{vc_tbl}{a table with the table, field, and vocabulary information for each
check}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{vc}}

\item{null_values}{a vector of NULL values (or other values that would not belong to accepted vocabularies
but are broadly accepted) that should be excluded when identifying non-valueset concepts}
}
\value{
a dataframe with summary information about each vocabulary that appears in the field,
with violations marked in a T/F field; vocabularies associated with the indicated
null_values are ignored, so proportions may not add up to 1 as a result
}
\description{
This function will use the \code{vocabulary.concept} table to identify the vocabulary
of each concept and determine how many rows comply with the standard vocabularies
expected for that field and how many rows violate these expectations.
}
