% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_vc_vs.R, R/valueset_conformance.R
\name{check_vs}
\alias{check_vs}
\title{Valueset Conformance}
\usage{
check_vs(
  vs_tbl,
  omop_or_pcornet = "omop",
  check_string = "vs",
  null_values = c(44814650L, 0L, 44814653L, 44814649L)
)

check_vs(
  vs_tbl,
  omop_or_pcornet = "omop",
  check_string = "vs",
  null_values = c(44814650L, 0L, 44814653L, 44814649L)
)
}
\arguments{
\item{vs_tbl}{a table with the table, field, and valueset information for each
check}

\item{omop_or_pcornet}{string indicating the CDM format of the data; defaults to \code{omop}}

\item{check_string}{an abbreviated identifier to identify all output from this module
defaults to \code{vs}}

\item{null_values}{a vector of NULL values (or other values that are not part of the valueset but are broadly accepted)
that should be excluded when identifying non-valueset concepts}

\item{valuesets}{a list that contains the following values:
\itemize{
\item name of list element: the check name identifier
\enumerate{
\item the name of the relevant valueset that appears in the specs directory
\item the field in which valueset violations should be identified
\item the name of the CDM table where the field is located
}
}}

\item{from_specs}{logical that determines if valueset exists in the specs directory}

\item{string_tbl_name}{the table name prefixed to the output}
}
\value{
a dataframe with summary information about each value that does not
appear in the valueset

a dataframe with summary information about each value that does not
appear in the valueset, or a row with dummy information if no violations are identified
}
\description{
Valueset Conformance

Valueset Conformance
}
