% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_check_metadata}
\alias{create_check_metadata}
\title{Create informational metadata file}
\usage{
create_check_metadata(
  check_tbls,
  metadata_file = NULL,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{check_tbls}{a list of table names from which the list of executed
checks should be extracted}

\item{metadata_file}{if one exists, a previously generated metadata file
to append to. the function will also highlight new
checks that will need to be described}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
a dataframe with all executed checks from the input tables

\if{html}{\out{<div class="sourceCode">}}\preformatted{     if a metadate file is provided, it will include all information from this
     file in addition to newly added checks that need additional metadata added

     if a metadata file is NOT provided, it will include all checks with some
     metadata pulled from the tables themselves and other fields left blank for the
     user to fill in
}\if{html}{\out{</div>}}
}
\description{
Create informational metadata file
}
