% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{process_bmc}
\alias{process_bmc}
\title{Best Mapped Concepts -- Processing}
\usage{
process_bmc(
  bmc_results,
  bmc_concepts_labelled,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{bmc_results}{the \code{bmc_counts} table output by check_bmc}

\item{bmc_concepts_labelled}{the \code{bmc_concepts} table output by check_bmc, with an additional
column called \code{include} added with "not best" or non-ideal concepts
marked with a 0 (optionally, "best" concepts can also be marked with a 1)}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
a dataframe summarizing the proportion of best vs not best concepts for a given check, indicated by
the appropriate value in the "include" column
}
\description{
Intakes the output of check_bmc in order to apply additional processing. This
includes applying the user-specified best/not best labels that were added to
the bmc_concepts output, then using those labels to compute proportions of
best vs not best concept representation in each check.
}
\examples{
# This function should be run after check_bmc has been executed for all
# network institutions and results have been combined into a common table

# All returned concepts should also be labelled to indicate whether they
# should be included as a "best" concept or not:

readr::read_csv(system.file('extdata', 'bmc_concept_examples.csv', package = 'ndq'))

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_bmc_process <- process_bmc(bmc_results = my_bmc_rslts,
                              bmc_concepts_labelled = my_bmc_concepts,
                              rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_bmc_process <- process_bmc(bmc_results = 'my_bmc_rslts',
                              bmc_concepts_labelled = 'my_bmc_concepts',
                              rslt_source = 'csv',
                              csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_bmc_process <- process_bmc(bmc_results = 'my_bmc_rslts',
                              bmc_concepts_labelled = 'my_bmc_concepts',
                              rslt_source = 'remote')
}

}
