% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mapped_concepts.R
\name{process_bmc}
\alias{process_bmc}
\title{Best Mapped Concepts -- Processing}
\usage{
process_bmc(bmc_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{bmc_results}{\emph{tabular input} || \strong{required}

The output of \code{check_bmc}. This table should include results for all
institutions that should be included in the computation of overall / "network level"
statistics.}

\item{rslt_source}{\emph{string} || defaults to \code{remote}

A string that identifies the location of the \code{bmc_results} table.
Acceptable values are
\itemize{
\item \code{local} - table is stored as a dataframe in the local R environment
\item \code{csv} - table is stored as a CSV file
\item \code{remote} - table is stored on a remote database
}}

\item{csv_rslt_path}{\emph{string} || defaults to \code{NULL}

If \code{rslt_source} has been set to \code{csv}, this parameter should indicate the path to
the result file(s). Otherwise, this parameter can be left as \code{NULL}}
}
\value{
A table summarizing the proportion of best vs not best
concepts for a given check, indicated by the user designation provided
in the \code{check_bmc} results
}
\description{
Intakes the output of \code{check_bmc} in order to apply additional processing. This
includes applying the user-specified best/not best labels that were added to
the \code{bmc_concepts} table, then using those labels to compute proportions of
best vs not best concept representation in each check.
}
\examples{
# This function should be run after check_bmc has been executed for all
# network institutions and results have been combined into a common table

## When results are kept locally:
\dontrun{
my_bmc_process <- process_bmc(bmc_results = my_bmc_rslts,
                              rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_bmc_process <- process_bmc(bmc_results = 'my_bmc_rslts',
                              rslt_source = 'csv',
                              csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_bmc_process <- process_bmc(bmc_results = 'my_bmc_rslts',
                              rslt_source = 'remote')
}

}
