% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_fact_documentation.R
\name{process_cfd}
\alias{process_cfd}
\title{Clinical Fact Documentation -- Processing}
\usage{
process_cfd(cfd_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{cfd_results}{\emph{tabular input} || \strong{required}

The tabular output of \code{check_cfd}. This table should include results for all
institutions that should be included in the computation of overall / "network level"
statistics.}

\item{rslt_source}{\emph{string} || defaults to \code{remote}

A string that identifies the location of the \code{cfd_results} table.
Acceptable values are
\itemize{
\item \code{local} - table is stored as a dataframe in the local R environment
\item \code{csv} - table is stored as a CSV file
\item \code{remote} - table is stored on a remote database
}}

\item{csv_rslt_path}{\emph{string} || defaults to \code{NULL}

If \code{rslt_source} has been set to \code{csv}, this parameter should indicate the path to
the result file(s). Otherwise, this parameter can be left as \code{NULL}}
}
\value{
This function will return all columns from the \code{cfd_results} input with
additional rows where site = \code{total} that reflect the combined results of
all institutions provided in the input data.
}
\description{
Intakes the combined output of \code{check_cfd} from all relevant institutions in order
to apply additional processing. It will compute overall counts/proportions across all
sites included in the input and tidy some of the descriptive metadata.
}
\examples{
# This function should be run after check_cfd has been executed for all
# network institutions and results have been combined into a common table

## When results are kept locally:
\dontrun{
my_cfd_process <- process_cfd(cfd_results = my_cfd_rslts,
                              rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_cfd_process <- process_cfd(cfd_results = 'my_cfd_rslts',
                              rslt_source = 'csv',
                              csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_cfd_process <- process_cfd(cfd_results = 'my_cfd_rslts',
                              rslt_source = 'remote')
}


}
