% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cycle_changes.R
\name{process_dc}
\alias{process_dc}
\title{Data Cycle Changes -- Processing}
\usage{
process_dc(
  dc_ct_results,
  dc_meta_results,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{dc_ct_results}{\emph{tabular input} || \strong{required}

The \code{dc_cts} output of \code{check_dc}. This table should include results for all
institutions that should be included in the computation of overall / "network level"
statistics.}

\item{dc_meta_results}{\emph{tabular input} || \strong{required}

The \code{dc_meta} output of \code{check_dc}}

\item{rslt_source}{\emph{string} || defaults to \code{remote}

A string that identifies the location of the \code{dc_results} table.
Acceptable values are
\itemize{
\item \code{local} - table is stored as a dataframe in the local R environment
\item \code{csv} - table is stored as a CSV file
\item \code{remote} - table is stored on a remote database
}}

\item{csv_rslt_path}{\emph{string} || defaults to \code{NULL}

If \code{rslt_source} has been set to \code{csv}, this parameter should indicate the path to
the result file(s). Otherwise, this parameter can be left as \code{NULL}}
}
\value{
This function will return the \code{dc_ct_results} table with overall / "network-wide" counts
and the proportion change between the two data models

The \code{dc_meta_results} table remains unchanged, so the original version returned by \code{check_dc}
can be used if needed
}
\description{
Intakes the output of \code{check_dc} in order to apply additional processing. This
includes computing percent change across data model versions and computing an
overall set of counts/percent changes across all sites included in the input.
}
\examples{
# This function should be run after check_dc has been executed for all
# network institutions and results have been combined into a common table

## When results are kept locally:
\dontrun{
my_dc_process <- process_dc(dc_ct_results = my_dc_cts,
                            dc_meta_results = my_dc_meta,
                            rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_dc_process <- process_dc(dc_ct_results = 'my_dc_cts',
                            dc_meta_results = 'my_dc_meta',
                            rslt_source = 'csv',
                            csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_dc_process <- process_dc(dc_ct_results = 'my_dc_cts',
                            dc_meta_results = 'my_dc_meta',
                            rslt_source = 'remote')
}


}
