% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cycle_changes.R
\name{process_dc}
\alias{process_dc}
\title{Data Cycle Changes -- Processing}
\usage{
process_dc(
  dc_ct_results,
  dc_meta_results,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{dc_ct_results}{the name of the table output by check_dc labelled \code{dc_cts} that contains
the previous and current data cycle counts}

\item{dc_meta_results}{the name of the table output by check_dc labelled \code{dc_meta} that contains
metadata associated with each executed check}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
dc_ct_results tbl with totals and additional change in proportion
and threshold columns; the dc_meta_results table remains unchanged, so the original
version should be used where needed
}
\description{
Intakes the output of check_dc in order to apply additional processing. This
includes computing percent change across data model versions and computing an
overall set of counts/percent changes across all sites included in the input.
}
\examples{
# This function should be run after check_dc has been executed for all
# network institutions and results have been combined into a common table

## When results are kept locally:
\dontrun{
my_dc_process <- process_dc(dc_ct_results = my_dc_cts,
                            dc_meta_results = my_dc_meta,
                            rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_dc_process <- process_dc(dc_ct_results = 'my_dc_cts',
                            dc_meta_results = 'my_dc_meta',
                            rslt_source = 'csv',
                            csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_dc_process <- process_dc(dc_ct_results = 'my_dc_cts',
                            dc_meta_results = 'my_dc_meta',
                            rslt_source = 'remote')
}


}
