% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_concordance.R
\name{process_dcon}
\alias{process_dcon}
\title{Domain Concordance -- Processing}
\usage{
process_dcon(dcon_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{dcon_results}{table output by check_dcon}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
dcon_tbl with additional columns with totals and proportions for the checks
}
\description{
Function to add a proportion column
when counts are computed for each domain and for combined
and to add total as a site summarizing cohorts from all sites
}
