% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_concordance.R
\name{process_dcon}
\alias{process_dcon}
\title{Domain Concordance -- Processing}
\usage{
process_dcon(dcon_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{dcon_results}{\emph{tabular input} || \strong{required}

The \code{dcon_counts} output of \code{check_dcon}. This table should include results for all
institutions that should be included in the computation of overall / "network level"
statistics.}

\item{rslt_source}{\emph{string} || defaults to \code{remote}

A string that identifies the location of the \code{dcon_results} table.
Acceptable values are
\itemize{
\item \code{local} - table is stored as a dataframe in the local R environment
\item \code{csv} - table is stored as a CSV file
\item \code{remote} - table is stored on a remote database
}}

\item{csv_rslt_path}{\emph{string} || defaults to \code{NULL}

If \code{rslt_source} has been set to \code{csv}, this parameter should indicate the path to
the result file(s). Otherwise, this parameter can be left as \code{NULL}}
}
\value{
This function will return a dataframe that, for each cohort pair, contains one row for
each cohort overlap computation type listed in the description,
including the associated raw counts and proportions
}
\description{
Intakes the output of \code{check_dcon} in order to apply additional processing. This
includes computing the following cohort overlap counts and proportions:
\itemize{
\item \code{cohort_1_only}: overall, patients in just cohort 1
\item \code{cohort_2_only}: overall, patients in just cohort 2
\item \code{combined}: overall, patients in both 1 and 2
\item \code{cohort_1_denom}: patients in cohort 1 not cohort 2
\item \code{cohort_2_denom}: patients in cohort 2 not cohort 1
\item \code{cohort_1_in_2}: patients in cohort 2 who are also in 1 (use cohort 2 as denominator)
\item \code{cohort_2_in_1}: patients in cohort 1 who are also in 2 (use cohort 1 as denominator)
}
}
\details{
Note that \code{cohort_1_in_2} and \code{cohort_2_in_1} will have the same raw count, but different
proportions since the denominator is different
}
\examples{
# This function should be run after check_dcon has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_dcon_process <- process_dcon(dcon_results = my_dcon_rslts,
                                rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_dcon_process <- process_dcon(dcon_results = 'my_dcon_rslts',
                                rslt_source = 'csv',
                                csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_dcon_process <- process_dcon(dcon_results = 'my_dcon_rslts',
                                rslt_source = 'remote')
}

}
