% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_concepts_present.R
\name{process_ecp}
\alias{process_ecp}
\title{Expected Concepts Present -- Processing}
\usage{
process_ecp(ecp_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{ecp_results}{table output by check_ecp}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
same input table with additional check_name_app column to indicate application level
}
\description{
Intakes the output of check_ecp in order to apply additional processing. This
includes creating a new check_name_app column to specify that the check
was computed at the person level.
}
\examples{
# This function should be run after check_ecp has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_ecp_process <- process_ecp(ecp_results = my_ecp_rslts,
                              rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_ecp_process <- process_ecp(ecp_results = 'my_ecp_rslts',
                              rslt_source = 'csv',
                              csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_ecp_process <- process_ecp(ecp_results = 'my_ecp_rslts',
                              rslt_source = 'remote')
}

}
