% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts_over_time.R
\name{process_fot}
\alias{process_fot}
\title{Facts Over Time -- Process}
\usage{
process_fot(
  fot_results,
  target_col = "row_cts",
  add_ratios = FALSE,
  ratio_mult = 10000,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{fot_results}{the table output by check_fot}

\item{target_col}{the numerical column in fot_results that should be used to
compute the heuristic; options are row_cts, row_pts, or row_visits}

\item{add_ratios}{boolean to indicate whether ratios / rates should be computed
if TRUE, the fot_results table should have an additional \code{total_pt}
column with counts for the denominator of the user's choosing

\if{html}{\out{<div class="sourceCode">}}\preformatted{              for example, the count of patients in that time period that
              had a visit
}\if{html}{\out{</div>}}}

\item{ratio_mult}{if add_ratios = TRUE, the numerical multiplier that should
be used to compute the rate}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
a list with two dataframes:
one with the summary heuristic (month / ((month-1)\emph{.25 +
(month+1)}.25 +
(month-12)*.5))
and another with summary values (mean, med, sd, q1, q3) based on
the heuristic
}
\description{
Facts Over Time -- Process
}
