% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_visit_id.R
\name{process_mf_visitid}
\alias{process_mf_visitid}
\title{Missing Field: Visit ID -- Processing}
\usage{
process_mf_visitid(
  mf_visitid_results,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{mf_visitid_results}{\emph{tabular input} || \strong{required}

The tabular output of \code{check_mf_visitid}. This table should include results for all
institutions that should be included in the computation of overall / "network level"
statistics.}

\item{rslt_source}{\emph{string} || defaults to \code{remote}

A string that identifies the location of the \code{mf_visitid_results} table.
Acceptable values are
\itemize{
\item \code{local} - table is stored as a dataframe in the local R environment
\item \code{csv} - table is stored as a CSV file
\item \code{remote} - table is stored on a remote database
}}

\item{csv_rslt_path}{\emph{string} || defaults to \code{NULL}

If \code{rslt_source} has been set to \code{csv}, this parameter should indicate the path to
the result file(s). Otherwise, this parameter can be left as \code{NULL}}
}
\value{
This function will return the \code{mf_visitid_results} table with the addition
of a column that reflects proportion of missing visits
}
\description{
Intakes the output of \code{check_mf_visitid} in order to apply additional processing.
This includes computing proportions of missing visits and computing overall
totals across all sites included in the input.
}
\examples{
# This function should be run after check_mf_visitid has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_mf_visitid_process <- process_mf_visitid(mf_visitid_results = my_mf_visitid_rslts,
                                            rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_mf_visitid_process <- process_mf_visitid(mf_visitid_results = 'my_mf_visitid_rslts',
                                            rslt_source = 'csv',
                                            csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_mf_visitid_process <- process_mf_visitid(mf_visitid_results = 'my_mf_visitid_rslts',
                                            rslt_source = 'remote')
}

}
