% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_visit_id.R
\name{process_mf_visitid}
\alias{process_mf_visitid}
\title{Missing Field: Visit ID -- Processing}
\usage{
process_mf_visitid(
  mf_visitid_results,
  rslt_source = "remote",
  csv_rslt_path = NULL
)
}
\arguments{
\item{mf_visitid_results}{output of check_mf_visitid}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
mf_visitid tbl with additional domain, total_ct, and proportion
column
}
\description{
Intakes the output of check_mf_visitid in order to apply additional processing.
This includes computing proportions of missing visits and computing overall
totals across all sites included in the input.
}
\examples{
# This function should be run after check_mf_visitid has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_mf_visitid_process <- process_mf_visitid(mf_visitid_results = my_mf_visitid_rslts,
                                            rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_mf_visitid_process <- process_mf_visitid(mf_visitid_results = 'my_mf_visitid_rslts',
                                            rslt_source = 'csv',
                                            csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_mf_visitid_process <- process_mf_visitid(mf_visitid_results = 'my_mf_visitid_rslts',
                                            rslt_source = 'remote')
}

}
