% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmapped_concepts.R
\name{process_uc}
\alias{process_uc}
\title{Unmapped Concepts -- Processing}
\usage{
process_uc(uc_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{uc_results}{\emph{tabular input} || \strong{required}

The tabular output of \code{check_uc}. This table should include results for all
institutions that should be included in the computation of overall / "network level"
statistics.}

\item{rslt_source}{\emph{string} || defaults to \code{remote}

A string that identifies the location of the \code{uc_results} table.
Acceptable values are
\itemize{
\item \code{local} - table is stored as a dataframe in the local R environment
\item \code{csv} - table is stored as a CSV file
\item \code{remote} - table is stored on a remote database
}}

\item{csv_rslt_path}{\emph{string} || defaults to \code{NULL}

If \code{rslt_source} has been set to \code{csv}, this parameter should indicate the path to
the result file(s). Otherwise, this parameter can be left as \code{NULL}}
}
\value{
This function will return all columns from the \code{uc_results} input with either an
additional column with unmapped proportions (by_year) or with additional rows that
include total unmapped counts/proportions across all sites (not by_year)
}
\description{
Intakes the output of \code{check_uc} in order to apply additional processing. This
includes either adding proportions (for longitudinal output) or computing overall
totals across all sites included in the input (for non-longitudinal output)
}
\examples{
# This function should be run after check_uc has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_uc_process <- process_uc(uc_results = my_uc_rslts,
                            rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_uc_process <- process_uc(uc_results = 'my_uc_rslts',
                            rslt_source = 'csv',
                            csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_uc_process <- process_uc(uc_results = 'my_uc_rslts',
                            rslt_source = 'remote')
}

}
