% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_conformance.R
\name{process_vc}
\alias{process_vc}
\title{Vocabulary Conformance -- Processing}
\usage{
process_vc(vc_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{vc_results}{the output of check_vc}

\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}
}
\value{
a list that contains two dataframes:
\itemize{
\item \code{vc_processed}: a dataframe with additional columns that include proportions of violations
and the overall summary
\item \code{vc_violations}: a dataframe with ONLY violating vocabularies
}
}
\description{
Intakes the output of check_vc in order to apply additional processing. This
includes computing row and patient proportions and computing overall totals
across all sites included in the input.
}
\examples{
# This function should be run after check_vc has been executed for all
# network institutions and results have been combined into a common table

# Once the labels have been applied, the function can be executed
## When results are kept locally:
\dontrun{
my_vc_process <- process_vc(vc_results = my_vc_rslts,
                            rslt_source = 'local')
}

## When results are kept in CSV files:
\dontrun{
my_vc_process <- process_vc(vc_results = 'my_vc_rslts',
                            rslt_source = 'csv',
                            csv_rslt_path = 'path/to/my/results')
}

## When results are kept on a remote database:
\dontrun{
my_vc_process <- process_vc(vc_results = 'my_vc_rslts',
                            rslt_source = 'remote')
}

}
