% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_conformance.R
\name{process_vc}
\alias{process_vc}
\title{Vocabulary Conformance -- Processing}
\usage{
process_vc(vc_results, rslt_source = "remote", csv_rslt_path = NULL)
}
\arguments{
\item{rslt_source}{the location of the results. acceptable values are \code{local} (stored as a dataframe in the R environment),
\code{csv} (stored as CSV files), or \code{remote} (stored on a remote DBMS); defaults to remote}

\item{csv_rslt_path}{if the results have been stored as CSV files, the path to the location
of these files. If the results are local or remote, leave NULL}

\item{vs_results}{the output of check_vs}
}
\value{
a list that contains two dataframes:
table with:
site
table_application
measurement_column
vocabulary_id
check_type
check_name
total_denom_ct
total_concept_ct
accepted_value
tot_ct: sum of total rows for given value in vocab
tot_prop: sum of proportion of all values for that vocab
tot_dist_concept_ct: total distinct concepts for that vocab
tot_dist_concept_prop: proportion of concepts in vocab of all concepts in table

\if{html}{\out{<div class="sourceCode">}}\preformatted{     and another table with a similar structure but only containing the vocabulary violations
}\if{html}{\out{</div>}}
}
\description{
Function to compute proportions of vocabulary values from vc check
}
