% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_n_summary.R
\name{summarize_large_n}
\alias{summarize_large_n}
\title{Add summary columns for a large N of sites}
\usage{
summarize_large_n(
  dq_output,
  num_col,
  grp_vars,
  check_string,
  time = FALSE,
  shape = "wide"
)
}
\arguments{
\item{dq_output}{\emph{tabular output} || \strong{required}

Any of the output tables that has been created by one of the \verb{check_*} functions
and has also passed through the associated \verb{process_*} function}

\item{num_col}{\emph{string} || \strong{required}

A string indicating the name of the numeric column in the \code{dq_output} table
that should be the basis for summary statistic computation}

\item{grp_vars}{\emph{string / vector} || \strong{required}

A string or vector listing the variables in \code{dq_output} that should be used
for grouping when comparing summary statistics.}

\item{check_string}{\emph{string} || \strong{required}

A string that matches the check string originally input into the analysis}

\item{time}{\emph{boolean} || defaults to \code{FALSE}

A boolean indicating whether the check is time dependent. This should only
apply for Facts Over Time output or Unmapped Concepts output in cases when it
was computed longitudinally}

\item{shape}{\emph{string} || defaults to \code{wide}

A string, either \code{long} or \code{wide}, indicating whether summary statistics should be
separate columns in the table (wide) or assigned to separate rows with the
name of the summary statistic in the site column (long)}
}
\value{
This function will return the original \code{dq_output} table with the addition of
summary statistics based on the numeric column provided. These will include:
\itemize{
\item If \code{time = TRUE}, the mean and median are returned.
\item If \code{time = FALSE}, the max, min, q1, q3, mean, and median are returned
}
}
\description{
For analyses that have a high volume of sites in the output, the standard, site-level
results may be difficult to visualize on a plot. This function will compute high level summary
statistics across all sites to provide another option for visualization. It will also
allow for users to compare the site level output to the overall / "network-level" statistics.
}
